+++ CHANGELOG.md
@@ -17,6 +17,12 @@ should change the heading of the (upcoming) version to include a major version b
 -->
 # 5.14.2
 
+## @rjsf/mui
+- Resolve the React error caused by the propagation of the `hideError` property to the DOM element, fixing [#3945](https://github.com/rjsf-team/react-jsonschema-form/issues/3945)
+
+## @rjsf/material-ui
+- Resolve the React error caused by the propagation of the `hideError` property to the DOM element, fixing [#3945](https://github.com/rjsf-team/react-jsonschema-form/issues/3945)
+
 ## @rjsf/utils
 
 - Update `sanitizeDataForNewSchema()` to avoid spreading strings and Arrays into the returned value when the old schema is of type `string` or `array` and the new schema is of type `object`. Fixing [#3922](https://github.com/rjsf-team/react-jsonschema-form/issues/3922)
@@ -32,7 +38,7 @@ should change the heading of the (upcoming) version to include a major version b
 
 - Added Fluent UI v9 (React Components) theme to playground
 - Update Fluent UI v9 and playground project references
-- Update eslint ignores to exclude new typescript build output folders 
+- Update eslint ignores to exclude new typescript build output folders
 
 # 5.14.0